+++ example/mxnet_face_detector_service_1/face_detector/retinaface/core/main/__init__.py
@@ -1,2 +1,0 @@
-# from .detector import FaceDetector
-# from .encoder import FaceEncoder