+++ example/mxnet_face_detector_service_1/face_detector/retinaface/core/main/detector_method/rcnn/symbol/__init__.py
@@ -1,4 +1,0 @@
-from .symbol_ssh import *
-from .symbol_mnet import *
-from .symbol_resnet import *
-from .symbol_common import *