+++ thrift4DL/server/__init__.py
@@ -1,1 +1,2 @@
 from .TModelPoolServer import TModelPoolServer
+from .base_handler import BatchingHandler