+++ thrift4DL/server/base_handler.py
@@ -1,17 +1,48 @@
 
-from .Thrift4DLService import ReceiverV2, DeliverV2
+from .Thrift4DLService import Receiver, Deliver
 import multiprocessing
 from .ttypes import TVisionResult
 from thrift.Thrift import TType, TMessageType, TApplicationException
 import traceback
 from queue import Empty
-from ..helpers import decode_image
 import numpy as np
 
 IDLE_QUEUE_BLOCK_TIME_SEC = 10
 
 
-class BaseHandlerV2(multiprocessing.Process):
+class BaseHandler():
+    def __init__(self, model_path, gpu_id, mem_fraction, client_queue, batch_group_timeout, batch_infer_size):
+        pass
+
+    def get_env(self, gpu_id, mem_fraction):
+        raise NotImplementedError
+
+    def get_model(self, model_path, env_params):
+        raise NotImplementedError
+
+    def preprocessing(self, model, input):
+        raise NotImplementedError
+
+    def postprocessing(self, model, input):
+        raise NotImplementedError
+
+    def predict(self, model, input):
+        raise NotImplementedError
+
+    def error_handle(self, args_dict):
+        raise NotImplementedError
+
+    def success_handle(self, result, args_dict):
+        raise NotImplementedError
+
+    def model_process(self, model, image_binary):
+        raise NotImplementedError
+
+    def run(self):
+        raise NotImplementedError
+
+
+class Handler(multiprocessing.Process):
     def __init__(self, model_path, gpu_id, mem_fraction, client_queue, batch_group_timeout, batch_infer_size):
         multiprocessing.Process.__init__(self)
         print("Init Handler")
@@ -21,9 +52,10 @@ class BaseHandlerV2(multiprocessing.Process):
         self.model_path = model_path
         self.batch_infer_size = batch_infer_size
         self.batch_group_timeout = self._milisec_to_sec(batch_group_timeout)
-        self.receiver = ReceiverV2(client_queue=self.client_queue)
-        self.deliver = DeliverV2()
+        self.receiver = Receiver()
+        self.deliver = Deliver()
         self._pid = np.random.randint(1000)
+        print(self.client_queue)
 
     def _milisec_to_sec(self, sec):
         return sec/1000
@@ -57,7 +89,6 @@ class BaseHandlerV2(multiprocessing.Process):
         return args_dict
 
     def model_process(self, model, image_binary):
-        # img_arr = decode_image(image_binary)
         img_arr = self.preprocessing(model, image_binary)
         pred_result = self.predict(model, img_arr)
         pred_result = self.postprocessing(model, pred_result)
@@ -86,40 +117,7 @@ class BaseHandlerV2(multiprocessing.Process):
             self.deliver.process(args_dict)
 
 
-class BatchingBaseHandlerV2(BaseHandlerV2):
-
-    def get_batch_v1(self):
-        """ Block queue for a while to wait incomming request
-        """
-        batch_input = []
-        is_done = False
-        is_empty = False
-        timeout = IDLE_QUEUE_BLOCK_TIME_SEC
-        while True:
-            try:
-                if is_done:
-                    # Reset state
-                    batch_input.clear()
-                    is_done = False
-                    is_empty = False
-                    timeout = IDLE_QUEUE_BLOCK_TIME_SEC
-                try:
-                    client = self.client_queue.get(block=True,
-                                                timeout=timeout)
-                    args_dict = self.receiver.process(client)
-                    batch_input.append(args_dict)
-                    timeout = self.batch_group_timeout
-                    # self.client_queue.task_done()
-                except Empty:
-                    is_empty = True
-
-                if (len(batch_input) >= self.batch_infer_size) or (is_empty and len(batch_input) > 0):
-                    is_done = True
-                    print(f"Process: {self._pid}:", len(batch_input))
-                    yield batch_input
-            except Exception as e:
-                print(traceback.format_exc())
-
+class BatchingHandler(Handler):
 
     def get_batch(self):
         """ Block queue for a while to wait incomming request
@@ -127,22 +125,17 @@ class BatchingBaseHandlerV2(BaseHandlerV2):
         batch_input = []
         is_done = False
         is_empty = False
-        is_block = True
         timeout = IDLE_QUEUE_BLOCK_TIME_SEC
         while True:
             try:
                 if is_done:
-                    # Reset state
                     batch_input.clear()
                     is_done = False
                     is_empty = False
-                    is_block = True
                     timeout = IDLE_QUEUE_BLOCK_TIME_SEC
-
-                print(is_block)
                 try:
-                    client = self.client_queue.get(block=is_block,
-                                                timeout=timeout)
+                    client = self.client_queue.get(block=True,
+                                                   timeout=timeout)
                     args_dict = self.receiver.process(client)
                     batch_input.append(args_dict)
                     timeout = self.batch_group_timeout
@@ -150,15 +143,10 @@ class BatchingBaseHandlerV2(BaseHandlerV2):
                 except Empty:
                     is_empty = True
 
-                if len(batch_input) > 0:
-                    is_block = False
-
-                
                 if (len(batch_input) >= self.batch_infer_size) or (is_empty and len(batch_input) > 0):
                     is_done = True
                     print(f"Process: {self._pid}:", len(batch_input))
                     yield batch_input
-
             except Exception as e:
                 print(traceback.format_exc())
 
@@ -166,6 +154,8 @@ class BatchingBaseHandlerV2(BaseHandlerV2):
         env_params = self.get_env(self.gpu_id, self.mem_fraction)
         model = self.get_model(self.model_path, env_params)
         for batch_input in self.get_batch():
+            import time
+
             if len(batch_input) > 0:
                 batch_image_binary = []
                 batch_pred_result = []
@@ -186,7 +176,8 @@ class BatchingBaseHandlerV2(BaseHandlerV2):
                     print(traceback.format_exc())
 
                 for pred_result in batch_pred_result:
-                    batch_final_result.append(self.postprocessing(model, pred_result))
+                    batch_final_result.append(
+                        self.postprocessing(model, pred_result))
 
                 for connection_info, pred_result in zip(batch_input, batch_final_result):
                     try:
@@ -203,37 +194,3 @@ class BatchingBaseHandlerV2(BaseHandlerV2):
                         self.deliver.process(connection_info)
                     except Exception as e:
                         print(traceback.format_exc())
-
-
-# class BatchingBaseHandlerV2(BatchingBaseHandlerV1):
-#     def get_batch(self):
-#         """ Block queue for a while to wait incomming request
-#         """
-#         batch_input = []
-#         is_done = False
-#         is_empty = False
-#         is_block = True
-#         timeout = IDLE_QUEUE_BLOCK_TIME_SEC
-#         while True:
-#             if is_done:
-#                 # Reset state
-#                 batch_input.clear()
-#                 is_done = False
-#                 is_empty = False
-#                 is_block = True
-#                 timeout = IDLE_QUEUE_BLOCK_TIME_SEC
-#             try:
-#                 client = self.client_queue.get(block=is_block,
-#                                                timeout=timeout)
-#                 args_dict = self.receiver.process(client)
-#                 batch_input.append(args_dict)
-#                 timeout = self.batch_group_timeout
-#                 is_block = False
-#                 self.client_queue.task_done()
-#             except Empty:
-#                 is_empty = True
-
-#             if (len(batch_input) >= self.batch_infer_size) or (is_empty and len(batch_input) > 0):
-#                 is_done = True
-#                 print(f"Process: {self._pid}:", len(batch_input))
-#                 yield batch_input